refactor: rename package from warehouse to carry and update SearchByLocality method to accept locality_id argument
