+++ test/carry/mock_carry_repository.go
@@ -1,4 +1,4 @@
-package warehouse
+package carry
 
 import (
 	"app/pkg/models"
@@ -21,6 +21,6 @@ func (m *MockCarryRepository) Create(v models.Carry) (w models.Carry, err error)
 }
 
 func (m *MockCarryRepository) SearchByLocality(locality_id int) (v map[string]models.CarryByLocality, err error) {
-	args := m.Called()
+	args := m.Called(locality_id)
 	return args.Get(0).(map[string]models.CarryByLocality), args.Error(1)
 }
